Rename source to src to match sample analyzer
